# Let erubi-rails loop more in each iteration