--- a/benchmarks/erubi-rails/benchmark.rb
+++ b/benchmarks/erubi-rails/benchmark.rb
@@ -12,8 +12,8 @@ EXPECTED_TEXT_SIZE = 9369
 app = Rails.application
 fake_controller = FakeDiscourseController.new
 
-run_benchmark(900) do
-  100.times do
+run_benchmark(10) do
+  10000.times do
     out = FakeDiscourseController.render :topics_show, assigns: fake_controller.stub_assigns
   end
 end